Move get_ray_num to utils
