+++ salsa/utils/__init__.py
@@ -4,6 +4,6 @@ Utility scripts
 
 """
 
-from salsa.utils.collect_files import check_rays, collect_files
+from salsa.utils.collect_files import check_rays, get_ray_num, collect_files
 
 from salsa.utils.functions import ion_p, ion_p_num
